+++ docs/changelog.rst
@@ -4,11 +4,13 @@ Changelog
 Version 0.1 (initial release)
 -----------------------------
 
+- :class:`jaxopt.AndersonAcceleration`
 - :class:`jaxopt.Bisection`
 - :class:`jaxopt.BlockCoordinateDescent`
 - :class:`jaxopt.GradientDescent`
 - :class:`jaxopt.MirrorDescent`
 - :class:`jaxopt.OptaxSolver`
+- :class:`jaxopt.PicardIteration`
 - :class:`jaxopt.PolyakSGD`
 - :class:`jaxopt.ProjectedGradient`
 - :class:`jaxopt.ProximalGradient`